LPS-166857 Moves LayoutsImporter and LayoutsImporterResultEntry to layout-api module
