+++ modules/apps/layout/layout-api/src/main/java/com/liferay/layout/importer/LayoutsImporterResultEntry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.page.template.importer;
+package com.liferay.layout.importer;
 
 /**
  * @author Rubén Pulido
